BUGFIX: Support for `null` and `""` as `sentinels` setting
If the sentinel setting exists it currently is always treated as a string or array containing sentinels
which causes errors if a setting overwritten in via env-var or unset by another settings.yaml.

This change adds support´ for null and empty string as allowed value for the key `sentinels`,
those values will yield an empty sentinels array and thus disable the use of sentinels.